week4/hmmspeak-flask/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai
webrtcvad==2.0.10